Fix clang errors and warnings. Minor improvements to surrounding code in ArAnalogGyro, fix potential bug in ArConfigArg::getString() (don't return reference to temporary std::string). Add debug flag to ArSocket.

+++ src/ArVCC4.cpp
@@ -1999,10 +1999,10 @@ bool ArVCC4::setDefaultRange()
   //    units = degrees / 0.1125
 
   // Set min tilt range
-  myPacket.byte4ToBuf(ArMath::roundInt(MIN_TILT/.1125) + 0x8000);
+  myPacket.byte4ToBuf(ArMath::roundInt((double)MIN_TILT/.1125) + 0x8000);
 
   // Set max tilt range
-  myPacket.byte4ToBuf(ArMath::roundInt(MAX_TILT/.1125) + 0x8000);
+  myPacket.byte4ToBuf(ArMath::roundInt((double)MAX_TILT/.1125) + 0x8000);
 
   requestBytes();
   return sendPacket(&myPacket);
